+++ librz/debug/meson.build
@@ -80,7 +80,6 @@ if get_option('use_gpl')
 endif
 
 if has_debugger
-  rz_debug_sources += ['p/debug_native.c']
 
   if host_machine.system() == 'linux' or host_machine.system() == 'android'
     rz_debug_sources += ['p/native/linux/linux_debug.c']
@@ -89,9 +88,98 @@ if has_debugger
     rz_debug_sources += ['p/native/linux/linux_coredump.c']
   endif
 
+  # Linux OS
+  if host_machine.system() == 'linux'
+    if host_machine.cpu_family() == 'x86_64' or host_machine.cpu_family() == 'x86'
+      rz_debug_sources += ['p/native/linux_x86_64.c']
+    elif host_machine.cpu_family() == 'arm'
+      rz_debug_sources += ['p/native/linux_arm.c']
+    elif host_machine.cpu_family() == 'aarch64'
+      rz_debug_sources += ['p/native/linux_arm64.c']
+    elif host_machine.cpu_family() == 'ppc' or host_machine.cpu_family() == 'ppc64'
+      rz_debug_sources += ['p/native/linux_ppc.c']
+    elif host_machine.cpu_family() == 'mips'
+      rz_debug_sources += ['p/native/linux_mips32.c']
+    elif host_machine.cpu_family() == 'mips64'
+      rz_debug_sources += ['p/native/linux_mips64.c']
+    else
+      rz_debug_sources += ['p/native/linux_other.c']
+    endif
+
+  # Windows OS
+  elif host_machine.system() == 'windows'
+    if host_machine.cpu_family() == 'x86' or host_machine.cpu_family() == 'x86_64' or host_machine.cpu_family() == 'arm' or host_machine.cpu_family() == 'aarch64'
+      rz_debug_sources += ['p/native/windows.c']
+    else
+      rz_debug_sources += ['p/native/unsupported.c']
+    endif
+
+  # FreeBSD OS
+  elif host_machine.system() == 'freebsd'
+    if host_machine.cpu_family() == 'x86' or host_machine.cpu_family() == 'x86_64' or host_machine.cpu_family() == 'aarch64'
+      rz_debug_sources += ['p/native/kfbsd.c']
+    else
+      rz_debug_sources += ['p/native/unsupported.c']
+    endif
+
+  # OpenBSD OS
+  elif host_machine.system() == 'openbsd'
+    if host_machine.cpu_family() == 'x86_64'
+      rz_debug_sources += ['p/native/openbsd.c']
+    else
+      rz_debug_sources += ['p/native/unsupported.c']
+    endif
+
+  # NetBSD OS
+  elif host_machine.system() == 'netbsd'
+    if host_machine.cpu_family() == 'x86' or host_machine.cpu_family() == 'x86_64'
+      rz_debug_sources += ['p/native/netbsd.c']
+    else
+      rz_debug_sources += ['p/native/unsupported.c']
+    endif
+
+  # DragonFlyBSD OS
+  elif host_machine.system() == 'dragonfly'
+    if host_machine.cpu_family() == 'x86_64'
+      rz_debug_sources += ['p/native/dragonfly.c']
+    else
+      rz_debug_sources += ['p/native/unsupported.c']
+    endif
+
+  # Darwin (macOS/iPadOS/iOS)
+  elif host_machine.system() == 'darwin'
+    if host_machine.cpu_family() == 'x86_64' or host_machine.cpu_family() == 'x86'
+      rz_debug_sources += ['p/native/apple_x86_64.c']
+    elif host_machine.cpu_family() == 'aarch64'
+      rz_debug_sources += ['p/native/apple_aarch64.c']
+    elif host_machine.cpu_family() == 'arm'
+      rz_debug_sources += ['p/native/apple_arm.c']
+    elif host_machine.cpu_family() == 'ppc'
+      rz_debug_sources += ['p/native/apple_ppc.c']
+    else
+      rz_debug_sources += ['p/native/unsupported.c']
+    endif
+
+  # Android
+  elif host_machine.system() == 'android'
+    if host_machine.cpu_family() == 'x86_64' or host_machine.cpu_family() == 'x86'
+      rz_debug_sources += ['p/native/android_x86_64.c']
+    elif host_machine.cpu_family() == 'arm'
+      rz_debug_sources += ['p/native/android_arm.c']
+    elif host_machine.cpu_family() == 'aarch64'
+      rz_debug_sources += ['p/native/android_arm64.c']
+    else
+      rz_debug_sources += ['p/native/unsupported.c']
+    endif
+
+  # Other OS
+  else
+    rz_debug_sources += ['p/native/unsupported.c']
+  endif
+
   if host_machine.system() != 'windows'
     rz_debug_sources += [
-      'p/native/procfs.c'
+      'p/native/procfs.c',
     ]
   endif
 
@@ -107,12 +195,17 @@ if has_debugger
 
   if host_machine.system() == 'freebsd'
     rz_debug_deps += cc.find_library('util', required: true, static: is_static_build)
+    rz_debug_sources += ['p/native/bsd/kfbsd_debug.c']
+  elif host_machine.system() == 'netbsd'
+    rz_debug_sources += ['p/native/bsd/netbsd_debug.c']
+  elif host_machine.system() == 'openbsd'
+    rz_debug_sources += ['p/native/bsd/openbsd_debug.c']
+  elif host_machine.system() == 'dragonfly'
+    rz_debug_sources += ['p/native/bsd/bsd_debug.c']
   endif
+
   if ['netbsd', 'openbsd', 'freebsd', 'dragonfly'].contains(host_machine.system())
     rz_debug_deps += cc.find_library('kvm', required: true, static: is_static_build)
-    rz_debug_sources += [
-      'p/native/bsd/bsd_debug.c',
-    ]
   endif
 endif
