+++ librz/debug/p/native/bsd/kfbsd_debug.c
@@ -21,14 +21,9 @@
 #include <limits.h>
 #include "bsd_debug.h"
 #include <rz_util/rz_log.h>
-#if __KFBSD__ || __DragonFly__
 #include <sys/user.h>
 #include <libutil.h>
-#elif __OpenBSD__ || __NetBSD__
-#include <sys/proc.h>
-#endif
 
-#if __KFBSD__
 static void addr_to_string(struct sockaddr_storage *ss, char *buffer, int buflen) {
 	char buffer2[INET6_ADDRSTRLEN];
 	struct sockaddr_in6 *sin6;
@@ -61,12 +56,9 @@ static void addr_to_string(struct sockaddr_storage *ss, char *buffer, int buflen
 			break;
 		}
 }
-#endif
 
 int bsd_handle_signals(RzDebug *dbg) {
-#if __KFBSD__ || __NetBSD__
 	siginfo_t siginfo;
-#if __KFBSD__
 	// Trying to figure out a bit by the signal
 	struct ptrace_lwpinfo linfo = { 0 };
 	int ret = ptrace(PT_LWPINFO, dbg->pid, (char *)&linfo, sizeof(linfo));
@@ -80,19 +72,6 @@ int bsd_handle_signals(RzDebug *dbg) {
 	}
 
 	siginfo = linfo.pl_siginfo;
-#else
-	struct ptrace_siginfo sinfo = { 0 };
-	if (ptrace(PT_GET_SIGINFO, dbg->pid, (char *)&sinfo, sizeof(sinfo)) == -1) {
-		if (errno == ESRCH) {
-			dbg->reason.type = RZ_DEBUG_REASON_DEAD;
-			return 0;
-		}
-		rz_sys_perror("ptrace PTRACE_GET_SIGINFO");
-		return -1;
-	}
-
-	siginfo = sinfo.psi_siginfo;
-#endif
 	dbg->reason.type = RZ_DEBUG_REASON_SIGNAL;
 	dbg->reason.signum = siginfo.si_signo;
 
@@ -111,9 +90,6 @@ int bsd_handle_signals(RzDebug *dbg) {
 	}
 
 	return 0;
-#else
-	return -1;
-#endif
 }
 
 int bsd_reg_write(RzDebug *dbg, int type, const ut8 *buf, int size) {
@@ -124,9 +100,7 @@ int bsd_reg_write(RzDebug *dbg, int type, const ut8 *buf, int size) {
 			(caddr_t)buf, sizeof(struct reg));
 		break;
 	case RZ_REG_TYPE_DRX:
-#if __KFBSD__ || __NetBSD__
 		r = ptrace(PT_SETDBREGS, dbg->pid, (caddr_t)buf, sizeof(struct dbreg));
-#endif
 		break;
 	case RZ_REG_TYPE_FPU:
 		r = ptrace(PT_SETFPREGS, dbg->pid, (caddr_t)buf, sizeof(struct fpreg));
@@ -137,17 +111,15 @@ int bsd_reg_write(RzDebug *dbg, int type, const ut8 *buf, int size) {
 }
 
 bool bsd_generate_corefile(RzDebug *dbg, char *path, RzBuffer *dest) {
-#if defined(__NetBSD__)
-	return ptrace(PT_DUMPCORE, dbg->pid, path, strlen(path)) != -1;
-#elif defined(__FreeBSD__) && __FreeBSD_version >= 1302000
+#if defined(__FreeBSD__) && __FreeBSD_version >= 1302000
 	struct ptrace_coredump pc = { .pc_fd = dest->fd, .pc_flags = PC_ALL, .pc_limit = 0 };
 	return ptrace(PT_COREDUMP, dbg->pid, (void *)&pc, sizeof(pc)) != -1;
 #else
 	return false;
 #endif
 }
+
 RzDebugInfo *bsd_info(RzDebug *dbg, const char *arg) {
-#if __KFBSD__
 	struct kinfo_proc *kp;
 	RzDebugInfo *rdi = RZ_NEW0(RzDebugInfo);
 	if (!rdi) {
@@ -188,129 +160,9 @@ RzDebugInfo *bsd_info(RzDebug *dbg, const char *arg) {
 	free(kp);
 
 	return rdi;
-#elif __OpenBSD__
-	struct kinfo_proc *kp;
-	char err[_POSIX2_LINE_MAX];
-	int rc;
-	RzDebugInfo *rdi = RZ_NEW0(RzDebugInfo);
-	if (!rdi) {
-		return NULL;
-	}
-
-	kvm_t *kd = kvm_openfiles(NULL, NULL, NULL, KVM_NO_FILES, err);
-	if (!kd) {
-		free(rdi);
-		return NULL;
-	}
-
-	kp = kvm_getprocs(kd, KERN_PROC_PID, dbg->pid, sizeof(*kp), &rc);
-	if (kp) {
-		rdi->pid = dbg->pid;
-		rdi->tid = dbg->tid;
-		rdi->uid = kp->p_uid;
-		rdi->gid = kp->p__pgid;
-		rdi->exe = rz_str_dup(kp->p_comm);
-
-		switch (kp->p_stat) {
-		case SDEAD:
-			rdi->status = RZ_DBG_PROC_DEAD;
-			break;
-		case SSTOP:
-			rdi->status = RZ_DBG_PROC_STOP;
-			break;
-		case SSLEEP:
-			rdi->status = RZ_DBG_PROC_SLEEP;
-			break;
-		default:
-			rdi->status = RZ_DBG_PROC_RUN;
-			break;
-		}
-	}
-
-	kvm_close(kd);
-
-	return rdi;
-#elif __NetBSD__
-	struct kinfo_proc2 *kp;
-	char err[_POSIX2_LINE_MAX];
-	int np;
-	RzDebugInfo *rdi = RZ_NEW0(RzDebugInfo);
-	if (!rdi) {
-		return NULL;
-	}
-
-	kvm_t *kd = kvm_openfiles(NULL, NULL, NULL, KVM_NO_FILES, err);
-	if (!kd) {
-		free(rdi);
-		return NULL;
-	}
-
-	kp = kvm_getproc2(kd, KERN_PROC_PID, dbg->pid, sizeof(*kp), &np);
-	if (kp) {
-		rdi->pid = dbg->pid;
-		rdi->tid = dbg->tid;
-		rdi->uid = kp->p_uid;
-		rdi->gid = kp->p__pgid;
-		rdi->exe = rz_str_dup(kp->p_comm);
-
-		rdi->status = RZ_DBG_PROC_STOP;
-
-		switch (kp->p_stat) {
-		case SDEAD:
-			rdi->status = RZ_DBG_PROC_DEAD;
-			break;
-		case SSTOP:
-			rdi->status = RZ_DBG_PROC_STOP;
-			break;
-		case SZOMB:
-			rdi->status = RZ_DBG_PROC_ZOMBIE;
-			break;
-		case SACTIVE:
-		case SIDL:
-		case SDYING:
-			rdi->status = RZ_DBG_PROC_RUN;
-			break;
-		default:
-			rdi->status = RZ_DBG_PROC_SLEEP;
-		}
-	}
-
-	kvm_close(kd);
-
-	return rdi;
-#endif
 }
 
 RzList *bsd_pid_list(RzDebug *dbg, int pid, RzList *list) {
-#if __KFBSD__
-#ifdef __NetBSD__
-#define KVM_OPEN_FLAG KVM_NO_FILES
-#define KVM_GETPROCS(kd, opt, arg, cntptr) \
-	kvm_getproc2(kd, opt, arg, sizeof(struct kinfo_proc2), cntptr)
-#define KP_COMM(x) (x)->p_comm
-#define KP_PID(x)  (x)->p_pid
-#define KP_PPID(x) (x)->p_ppid
-#define KP_UID(x)  (x)->p_uid
-#define KINFO_PROC kinfo_proc2
-#elif defined(__OpenBSD__)
-#define KVM_OPEN_FLAG KVM_NO_FILES
-#define KVM_GETPROCS(kd, opt, arg, cntptr) \
-	kvm_getprocs(kd, opt, arg, sizeof(struct kinfo_proc), cntptr)
-#define KP_COMM(x) (x)->p_comm
-#define KP_PID(x)  (x)->p_pid
-#define KP_PPID(x) (x)->p_ppid
-#define KP_UID(x)  (x)->p_uid
-#define KINFO_PROC kinfo_proc
-#elif __DragonFly__
-#define KVM_OPEN_FLAG O_RDONLY
-#define KVM_GETPROCS(kd, opt, arg, cntptr) \
-	kvm_getprocs(kd, opt, arg, cntptr)
-#define KP_COMM(x) (x)->kp_comm
-#define KP_PID(x)  (x)->kp_pid
-#define KP_PPID(x) (x)->kp_ppid
-#define KP_UID(x)  (x)->kp_uid
-#define KINFO_PROC kinfo_proc
-#else
 #define KVM_OPEN_FLAG O_RDONLY
 #define KVM_GETPROCS(kd, opt, arg, cntptr) \
 	kvm_getprocs(kd, opt, arg, cntptr)
@@ -319,7 +171,6 @@ RzList *bsd_pid_list(RzDebug *dbg, int pid, RzList *list) {
 #define KP_PPID(x) (x)->ki_ppid
 #define KP_UID(x)  (x)->ki_uid
 #define KINFO_PROC kinfo_proc
-#endif
 	char errbuf[_POSIX2_LINE_MAX];
 	struct KINFO_PROC *kp, *entry;
 	int cnt = 0;
@@ -349,12 +200,10 @@ RzList *bsd_pid_list(RzDebug *dbg, int pid, RzList *list) {
 	}
 
 	kvm_close(kd);
-#endif
 	return list;
 }
 
 RzList *bsd_native_sysctl_map(RzDebug *dbg) {
-#if __KFBSD__
 	int mib[4];
 	size_t len = 0;
 	char *buf = NULL;
@@ -419,74 +268,9 @@ RzList *bsd_native_sysctl_map(RzDebug *dbg) {
 	}
 	free(buf);
 	return list;
-#elif __OpenBSD__
-	int mib[3];
-	size_t len;
-	struct kinfo_vmentry entry;
-	u_long old_end = 0;
-	RzList *list = NULL;
-	RzDebugMap *map;
-	u_long map_start = 0;
-	u_long map_end = 0;
-	ut64 offset = 0;
-	int perm = 0;
-
-	len = sizeof(entry);
-	mib[0] = CTL_KERN;
-	mib[1] = KERN_PROC_VMMAP;
-	mib[2] = dbg->pid;
-	entry.kve_start = 0;
-
-	if (sysctl(mib, 3, &entry, &len, NULL, 0) == -1) {
-		RZ_LOG_ERROR("Could not get memory map: %s\n", strerror(errno));
-		return NULL;
-	}
-
-	list = rz_debug_map_list_new();
-	if (!list)
-		return NULL;
-
-	while (sysctl(mib, 3, &entry, &len, NULL, 0) != -1) {
-		if (old_end == entry.kve_end) {
-			/* No more entries */
-			break;
-		}
-		/* path to vm obj is not included in kinfo_vmentry.
-		 * see usr.sbin/procmap for namei-cache lookup.
-		 */
-		map_start = entry.kve_start;
-		map_end = entry.kve_end;
-
-		if (entry.kve_protection & KVE_PROT_READ) {
-			perm |= RZ_PERM_R;
-		}
-		if (entry.kve_protection & KVE_PROT_WRITE) {
-			perm |= RZ_PERM_W;
-		}
-		if (entry.kve_protection & KVE_PROT_EXEC) {
-			perm |= RZ_PERM_X;
-		}
-
-		map = rz_debug_map_new("", map_start, map_end,
-			perm, 0);
-		if (!map) {
-			break;
-		}
-		map->offset = entry.kve_offset;
-		rz_list_append(list, map);
-
-		entry.kve_start = entry.kve_start + 1;
-		old_end = entry.kve_end;
-	}
-
-	return list;
-#else
-	return NULL;
-#endif
 }
 
 RzList *bsd_desc_list(int pid) {
-#if __KFBSD__
 	RzList *ret = NULL;
 	int perm, type, mib[4];
 	size_t len;
@@ -586,131 +370,8 @@ RzList *bsd_desc_list(int pid) {
 
 	free(buf);
 	return ret;
-#elif __NetBSD__
-	RzList *ret = NULL;
-	char path[512], file[512], buf[512];
-	struct dirent *de;
-	RzDebugDesc *desc;
-	int type, perm;
-	int len, len2;
-	struct stat st;
-	DIR *dd = NULL;
-
-	rz_strf(path, "/proc/%i/fd/", pid);
-	if (!(dd = opendir(path))) {
-		rz_sys_perror("opendir /proc/x/fd");
-		return NULL;
-	}
-	ret = rz_list_newf((RzListFree)rz_debug_desc_free);
-	if (!ret) {
-		closedir(dd);
-		return NULL;
-	}
-	while ((de = (struct dirent *)readdir(dd))) {
-		if (de->d_name[0] == '.') {
-			continue;
-		}
-		len = strlen(path);
-		len2 = strlen(de->d_name);
-		if (len + len2 + 1 >= sizeof(file)) {
-			RZ_LOG_ERROR("Filename is too long.\n");
-			goto fail;
-		}
-		memcpy(file, path, len);
-		memcpy(file + len, de->d_name, len2 + 1);
-		buf[0] = 0;
-		if (readlink(file, buf, sizeof(buf) - 1) == -1) {
-			RZ_LOG_ERROR("readlink %s failed.\n", file);
-			goto fail;
-		}
-		buf[sizeof(buf) - 1] = 0;
-		type = perm = 0;
-		if (stat(file, &st) != -1) {
-			type = st.st_mode & S_IFIFO ? 'P' : st.st_mode & S_IFSOCK ? 'S'
-				: st.st_mode & S_IFCHR                            ? 'C'
-										  : '-';
-		}
-		if (lstat(path, &st) != -1) {
-			if (st.st_mode & S_IRUSR) {
-				perm |= RZ_PERM_R;
-			}
-			if (st.st_mode & S_IWUSR) {
-				perm |= RZ_PERM_W;
-			}
-		}
-		// TODO: Offset
-		desc = rz_debug_desc_new(atoi(de->d_name), buf, perm, type, 0);
-		if (!desc) {
-			break;
-		}
-		rz_list_append(ret, desc);
-	}
-	closedir(dd);
-	return ret;
-
-fail:
-	rz_list_free(ret);
-	closedir(dd);
-	return NULL;
-#else
-	return NULL;
-#endif
 }
 
-#if __OpenBSD__
-static int get_rz_status(int stat) {
-	switch (stat) {
-	case SRUN:
-	case SIDL:
-	case SONPROC:
-		return RZ_DBG_PROC_RUN;
-	case SSTOP:
-		return RZ_DBG_PROC_STOP;
-	case SZOMB:
-		return RZ_DBG_PROC_ZOMBIE;
-	case SSLEEP:
-		return RZ_DBG_PROC_SLEEP;
-	default:
-		return RZ_DBG_PROC_DEAD;
-	}
-}
-
-static RzList *openbsd_thread_list(RzDebug *dbg, int pid, RzList *list) {
-	int mib[4] = { CTL_KERN, KERN_PROC, KERN_PROC_PID, pid };
-	struct kinfo_proc *kp;
-	size_t len = 0;
-	size_t max;
-	int i = 0;
-
-	if (sysctl(mib, 4, NULL, &len, NULL, 0) == -1) {
-		rz_list_free(list);
-		return NULL;
-	}
-
-	len += sizeof(*kp) + len / 10;
-	kp = malloc(len);
-	if (sysctl(mib, 4, kp, &len, NULL, 0) == -1) {
-		free(kp);
-		rz_list_free(list);
-		return NULL;
-	}
-
-	max = len / sizeof(*kp);
-	for (i = 0; i < max; i++) {
-		int pid_stat = get_rz_status(kp[i].p_stat);
-
-		RzDebugPid *pid_info = rz_debug_pid_new(kp[i].p_comm, kp[i].p_tid,
-			kp[i].p_uid, pid_stat, (ut64)kp[i].p_wchan);
-
-		rz_list_append(list, pid_info);
-	}
-
-	free(kp);
-	return list;
-}
-#endif
-
-#if __KFBSD__
 static int get_rz_status(int stat) {
 	switch (stat) {
 	case SRUN:
@@ -762,86 +423,8 @@ static RzList *kfbsd_thread_list(RzDebug *dbg, int pid, RzList *list) {
 	free(kp);
 	return list;
 }
-#endif
-
-#if __NetBSD__
-static int get_rz_status(int stat) {
-	switch (stat) {
-	case LSRUN:
-	case LSONPROC:
-	case LSIDL:
-		return RZ_DBG_PROC_RUN;
-	case LSSTOP:
-		return RZ_DBG_PROC_STOP;
-	case LSZOMB:
-		return RZ_DBG_PROC_ZOMBIE;
-	case LSSLEEP:
-		return RZ_DBG_PROC_SLEEP;
-	case LSSUSPENDED:
-		return RZ_DBG_PROC_STOP;
-	default:
-		return RZ_DBG_PROC_DEAD;
-	}
-}
-
-static RzList *netbsd_thread_list(RzDebug *dbg, int pid, RzList *list) {
-	int mib[6] = { CTL_KERN, KERN_PROC2, KERN_PROC_PID, pid, sizeof(struct kinfo_proc2), 0 };
-	struct kinfo_proc2 *kp;
-	size_t len = 0;
-	size_t max;
-	int i = 0;
-
-	if (sysctl(mib, 6, NULL, &len, NULL, 0) == -1) {
-		rz_list_free(list);
-		return NULL;
-	}
-
-	len += sizeof(*kp) + len / 10;
-	kp = malloc(len);
-	if (!kp) {
-		rz_list_free(list);
-		return NULL;
-	}
-
-	mib[5] = len / sizeof(struct kinfo_proc2);
-
-	if (sysctl(mib, 6, kp, &len, NULL, 0) == -1) {
-		free(kp);
-		rz_list_free(list);
-		return NULL;
-	}
-
-	max = len / sizeof(*kp);
-
-	for (i = 0; i < max; i++) {
-		int pid_stat = get_rz_status(kp[i].p_stat);
-
-		RzDebugPid *pid_info = rz_debug_pid_new(kp[i].p_comm, kp[i].p_pid,
-			kp[i].p_uid, pid_stat, (ut64)kp[i].p_wchan);
-
-		if (pid_info) {
-			rz_list_append(list, pid_info);
-		}
-	}
-
-	free(kp);
-	return list;
-}
-#endif
 
 RzList *bsd_thread_list(RzDebug *dbg, int pid, RzList *list) {
-#if __KFBSD__
 	RzList *thread_list = kfbsd_thread_list(dbg, pid, list);
 	return thread_list;
-#elif __NetBSD__
-	RzList *thread_list = netbsd_thread_list(dbg, pid, list);
-	return thread_list;
-#elif __OpenBSD__
-	RzList *thread_list = openbsd_thread_list(dbg, pid, list);
-	return thread_list;
-#else
-	RZ_LOG_ERROR("bsd_thread_list unsupported on this platform\n");
-	rz_list_free(list);
-	return NULL;
-#endif
 }
