+++ librz/debug/p/native/linux/linux_coredump.h
@@ -60,8 +60,6 @@
 
 #define SIZE_PR_FNAME 16
 
-#define RZ_DEBUG_REG_T struct user_regs_struct
-
 #define SIZE_NT_FILE_DESCSZ sizeof(unsigned long) * 3 /* start_address * end_address * offset_address */
 /*
 NT_FILE layout:
