+++ librz/debug/p/native/reg/linux-ppc64.h
