+++ test/db/archos/linux-ppc/dbg_sig
@@ -1,5 +1,6 @@
 NAME=Signal handling ppc64le
 FILE=bins/elf/ppc64le/dbg_signal_test
+BROKEN=1
 ARGS=-a ppc -d
 CMDS=<<EOF
 dkl~SIGTERM
